ai-incident-analyzer/services/log-processor/requirements.txt
kafka-python==2.0.2
httpx==0.26.0
